Update Accumulate Date Plan and Result

+++ WebApplication1/WebApplication1/Models/FTDenpyo.cs
@@ -29,11 +29,11 @@ namespace WebApplication1.Models
         public int A2_Lot { get; set; }
         public int A3_Lot { get; set; }
         public int A4_Lot { get; set; }
-        public int Plan_today { get; set; }
-        public int Result_today { get; set; }
-        public int Calulate_today { get; set; }
-        public int Plan_yesterday { get; set; }
-        public int Result_yesterday { get; set; }
-        public int Calulate_yesterday { get; set; }
+        public float Plan_today { get; set; }
+        public float Result_today { get; set; }
+        public float Calulate_today { get; set; }
+        public float Plan_yesterday { get; set; }
+        public float Result_yesterday { get; set; }
+        public float Calulate_yesterday { get; set; }
     }
 }
